requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
tweepy>=4.14.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
grad-cam>=1.4.8
